requirements.txt
Django==3.1.3
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.1.0
whitenoise==5.2.0
python-dotenv==0.19.2